+++ lightning/src/crypto/utils.rs
